+++ NEWS.md
@@ -1,3 +1,5 @@
+# themis (development version)
+
 # themis 0.1.4
 
 * export `smote()`, `adasyn()`, and `bsmote()` functions to users.