+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `step_rose()` now correctly allows you to use characters variables. (#26)
 * `step_tomek()` now ignore non-predictor variables when appropriate. (#51)
+* Update examples to no longer use `iris` or `okc` data sets.
 
 # themis 0.1.4