+++ NEWS.md
@@ -1,4 +1,4 @@
-# themis (development version)
+# themis 0.2.0
 
 ## New steps